requirements.txt
google-adk==1.14.0
google-cloud-logging==3.11.0
google-cloud-bigquery==3.25.0
google-auth==2.35.0
langchain-community==0.3.27
wikipedia==1.4.0
fastapi
uvicorn
python-dotenv==1.0.1
requests